pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project]
name = "european-power-fair-value"
version = "0.1.0"
description = "DE/LU hourly power price — data, QA, forecast, plots"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.1",
    "numpy>=1.26",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "pytz>=2024.1",
    "matplotlib>=3.8",
    "scikit-learn>=1.4",
    "openai>=1.0",
    "pyyaml>=6.0",
    "pyarrow>=15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.3",
]

[tool.ruff]
line-length = 100
target-version = "py311"
